Se agregó la capa de servicio y capa de controladores.

+++ as2final-backend/src/main/java/gt/edu/umg/as2final/model/SpecializationEntity.java
@@ -26,7 +26,7 @@ public class SpecializationEntity implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     
     // Nombre de la especialización
     @Size(max = 50)
@@ -44,15 +44,15 @@ public class SpecializationEntity implements Serializable {
     public SpecializationEntity() {
     }
 
-    public SpecializationEntity(Integer id) {
+    public SpecializationEntity(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
